requirements.txt
numpy==1.26.4          
pandas==2.2.3          
scikit-learn==1.5.1    
matplotlib==3.10.0
seaborn==0.13.2
imbalanced-learn==0.12.3 
joblib==1.4.2
pyyaml==6.0.2
pytest==8.3.4
python-dotenv==1.1.0
imbalanced-learn==0.14.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6